main.py
import os
from dotenv import load_dotenv
from google import genai
from google.genai import types
import argparse

load_dotenv()
api_key = os.environ.get("GEMINI_API_KEY")

if api_key is None:
    raise RuntimeError("No API key set.")

client = genai.Client(api_key=api_key)

parser = argparse.ArgumentParser(description="AI Agent Chatbot")
parser.add_argument("user_prompt", type=str, help="User prompt")
parser.add_argument("--verbose", action="store_true", help="Enable verbose output")
args = parser.parse_args()

messages = [types.Content(role="user", parts=[types.Part(text=args.user_prompt)])]

response = client.models.generate_content(
    model='gemini-2.5-flash', contents=messages
)

if response.usage_metadata is None:
    raise RuntimeError("No response returned. Possible API request failure.")

if args.verbose is True:
    print(f"User prompt: {args.user_prompt}")
    print(f"Prompt tokens: {response.usage_metadata.prompt_token_count}")
    print(f"Response tokens: {response.usage_metadata.candidates_token_count}")

print(response.text)

def main():
    print("Hello from ai-agent-bootdev!")


if __name__ == "__main__":
    main()



messages = [types.Content(role="user", parts=[types.Part(text=args.user_prompt)])]